refactor: posts index now using components
