+++ src/styles/components/pagination/BasicPagination.module.scss
@@ -1,0 +1,29 @@
+.pagination {
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  gap: 3px;
+
+  &__number {
+    background: $darker-white;
+    color: $black;
+    border: 3px solid $black;
+    font-family: $main-font;
+    padding: 5px;
+    font-size: 1.1rem;
+    font-weight: 700;
+    cursor: pointer;
+    transition: 0.2s ease-in-out;
+
+    &--active {
+      background: $black;
+      color: $darker-white;
+      pointer-events: none;
+    }
+
+    &:hover {
+      background: $black;
+      color: $darker-white;
+    }
+  }
+}
