+++ src/styles/pages/posts/index.module.scss
@@ -102,34 +102,4 @@
       }
     }
   }
-
-  &__pagination {
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    gap: 3px;
-
-    &-number {
-      background: $darker-white;
-      color: $black;
-      border: 3px solid $black;
-      font-family: $main-font;
-      padding: 5px;
-      font-size: 1.1rem;
-      font-weight: 700;
-      cursor: pointer;
-      transition: 0.2s ease-in-out;
-
-      &--active {
-        background: $black;
-        color: $darker-white;
-        pointer-events: none;
-      }
-
-      &:hover {
-        background: $black;
-        color: $darker-white;
-      }
-    }
-  }
 }
